Simulated the synchrotron topup PVs and had them "link" #22
Simulated the current loss and current gain of the synchrotron storage ring. The current loss and gain increment are calculated so they will take 10 minutes and 15 seconds respectively. The countdown is also simulated: based on the change of the current (this will only really work as a method if the rate of change is constant). Also wrote tests.